src/chatbot_builder/run.py
#!/usr/bin/env python
"""
Flask API Server Runner
Usage: python run.py
"""

import os
from dotenv import load_dotenv
from app import app, db

load_dotenv()

if __name__ == '__main__':
    with app.app_context():
        db.create_all()
    
    port = int(os.environ.get('PORT', 5000))
    debug = os.environ.get('FLASK_ENV') == 'development'
    
    print(f"Starting Flask API on port {port}...")
    print(f"Debug mode: {debug}")
    print(f"Database: {app.config['SQLALCHEMY_DATABASE_URI']}")
    
    app.run(
        host='0.0.0.0',
        port=port,
        debug=debug,
        use_reloader=debug
    )
