requirements.txt
# Core ML/AI
tensorflow==2.13.0
tensorflow-lite==2.13.0
numpy>=1.21.0,<2.0.0
pandas>=1.3.0
scikit-learn>=1.0.0

# NLP
nltk>=3.7
spacy>=3.5.0

# Web Framework
flask>=2.2.0
flask-cors>=3.0.10
werkzeug>=2.2.0

# Utilities
python-dotenv>=0.19.0
pydantic>=1.9.0
pyyaml>=6.0

# Serialization
protobuf>=3.20.0,<5.0.0
